+++ extensions/api/data-management/policydefinition/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/policy/service/PolicyServiceImpl.java
@@ -15,9 +15,9 @@
 package org.eclipse.dataspaceconnector.api.datamanagement.policy.service;
 
 import org.eclipse.dataspaceconnector.api.result.ServiceResult;
-import org.eclipse.dataspaceconnector.policy.model.Policy;
+import org.eclipse.dataspaceconnector.policy.model.PolicyDefinition;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
-import org.eclipse.dataspaceconnector.spi.policy.store.PolicyStore;
+import org.eclipse.dataspaceconnector.spi.policy.store.PolicyDefinitionStore;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
 import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
 import org.jetbrains.annotations.NotNull;
@@ -30,29 +30,29 @@ import static java.util.stream.Collectors.toList;
 public class PolicyServiceImpl implements PolicyService {
 
     private final TransactionContext transactionContext;
-    private final PolicyStore policyStore;
+    private final PolicyDefinitionStore policyStore;
     private final ContractDefinitionStore contractDefinitionStore;
 
-    public PolicyServiceImpl(TransactionContext transactionContext, PolicyStore policyStore, ContractDefinitionStore contractDefinitionStore) {
+    public PolicyServiceImpl(TransactionContext transactionContext, PolicyDefinitionStore policyStore, ContractDefinitionStore contractDefinitionStore) {
         this.transactionContext = transactionContext;
         this.policyStore = policyStore;
         this.contractDefinitionStore = contractDefinitionStore;
     }
 
     @Override
-    public Policy findById(String policyId) {
+    public PolicyDefinition findById(String policyId) {
         return transactionContext.execute(() ->
                 policyStore.findById(policyId));
     }
 
     @Override
-    public @NotNull Collection<Policy> query(QuerySpec query) {
+    public @NotNull Collection<PolicyDefinition> query(QuerySpec query) {
         return transactionContext.execute(() ->
                 policyStore.findAll(query).collect(toList()));
     }
 
     @Override
-    public @NotNull ServiceResult<Policy> deleteById(String policyId) {
+    public @NotNull ServiceResult<PolicyDefinition> deleteById(String policyId) {
 
         var contractFilter = format("contractPolicyId = %s ", policyId);
         var accessFilter = format("accessPolicyId = %s ", policyId);
@@ -85,7 +85,7 @@ public class PolicyServiceImpl implements PolicyService {
     }
 
     @Override
-    public @NotNull ServiceResult<Policy> create(Policy policy) {
+    public @NotNull ServiceResult<PolicyDefinition> create(PolicyDefinition policy) {
 
         return transactionContext.execute(() -> {
             if (policyStore.findById(policy.getUid()) == null) {